refactor app graphql and functions
