+++ app/kudoku/functions_/getKudosInfo.ts
