+++ app/kudoku/graphql_/query/getAllCashAccount.ts
