+++ app/kudoku/graphql_/query/getAllDebitAccount.ts
