+++ app/kudoku/graphql_/query/getAllEMoneyAccount.ts
