+++ app/kudoku/graphql_/query/getAllEWalletAccount.ts
