+++ app/kudoku/transaction/[accountType]/[id]/functions_/index.ts
@@ -1,0 +1,1 @@
+export * from './fetchAllAccounts';
