+++ app/kudoku/transaction/empty/graphql_/query/getAllCashAccount.ts
