+++ app/kudoku/transaction/empty/graphql_/query/getAllDebitAccount.ts
