+++ app/kudoku/transaction/empty/graphql_/query/getAllEMoneyAccount.ts
