+++ app/kudoku/transaction/empty/graphql_/query/getAllEWalletAccount.ts
