+++ app/kudoku/transaction/empty/graphql_/query/index.ts
