+++ app/kudoku/transaction/graphql_/query/getAllCashAccount.ts
