+++ app/kudoku/transaction/graphql_/query/getAllDebitAccount.ts
