+++ app/kudoku/transaction/graphql_/query/getAllEMoneyAccount.ts
