+++ app/kudoku/transaction/graphql_/query/getAllEWalletAccount.ts
