+++ app/kudoku/transaction/page.tsx
@@ -6,7 +6,7 @@ import {
   getAllEMoneyAccount,
   getAllEWalletAccount,
   getAllPayLaterAccount,
-} from './_graphql/query';
+} from './graphql_/query';
 
 export default async function Page() {
   const nextCookies = cookies();
