+++ lib/EmptyAccount/index.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import EmptyTransaction from './index';
+import EmptyAccount from './index';
 
 // const ThemeDecorator = (Story: any) => (
 //   <ThemeContextProvider>
@@ -9,20 +9,20 @@ import EmptyTransaction from './index';
 
 export default {
   title: 'Lib/Empty Transaction',
-  component: EmptyTransaction,
+  component: EmptyAccount,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
   //   decorators: [ThemeDecorator],
-} as ComponentMeta<typeof EmptyTransaction>;
+} as ComponentMeta<typeof EmptyAccount>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof EmptyTransaction> = () => {
+const Template: ComponentStory<typeof EmptyAccount> = () => {
   const token =
     'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NDA0ODIxZGMzNjM1OGNhZDU5NDA2MzEiLCJpYXQiOjE2Nzk5MDc0MDV9.ESWTAk0dY_N4F1HwcW6LIhzzNi17XTyVGGXM-au0ank';
 
   return (
     <>
-      <EmptyTransaction token={token} />
+      <EmptyAccount token={token} />
     </>
   );
 };
